Stub out tests for URLSession + URLSessionProtocol.
