+++ Sources/PaymentsCore/Networking/APIClient.swift
@@ -67,30 +67,3 @@ class APIClientDecoder {
         }
     }
 }
-
-protocol URLSessionProtocol {
-    func performRequest(with urlRequest: URLRequest) async throws -> (Data, URLResponse)
-}
-
-extension URLSession: URLSessionProtocol {
-
-    func performRequest(with urlRequest: URLRequest) async throws -> (Data, URLResponse) {
-        try await data(for: urlRequest)
-    }
-
-    @available(iOS, deprecated: 15.0, message: "This extension is no longer necessary. Use API built into SDK")
-    func data(for request: URLRequest) async throws -> (Data, URLResponse) {
-        try await withCheckedThrowingContinuation { continuation in
-            let task = self.dataTask(with: request) { data, response, error in
-                guard let data = data, let response = response else {
-                    let error = error ?? URLError(.badServerResponse)
-                    return continuation.resume(throwing: error)
-                }
-
-                continuation.resume(returning: (data, response))
-            }
-
-            task.resume()
-        }
-    }
-}
