+++ Sources/PaymentsCore/Networking/URLSession+URLSessionProtocol.swift
@@ -1,0 +1,30 @@
+//
+//  URLSession+URLSessionProtocol.swift
+//  PaymentsCore
+//
+//  Created by Steven Shropshire on 2/8/22.
+//
+
+import Foundation
+
+extension URLSession: URLSessionProtocol {
+
+    func performRequest(with urlRequest: URLRequest) async throws -> (Data, URLResponse) {
+        if #available(iOS 15, *) {
+            return try await self.data(for: urlRequest)
+        } else {
+            return try await withCheckedThrowingContinuation { continuation in
+                let task = self.dataTask(with: urlRequest) { data, response, error in
+                    guard let data = data, let response = response else {
+                        let error = error ?? URLError(.badServerResponse)
+                        return continuation.resume(throwing: error)
+                    }
+
+                    continuation.resume(returning: (data, response))
+                }
+
+                task.resume()
+            }
+        }
+    }
+}
