+++ UnitTests/PaymentsCoreTests/URLSession+URLSessionProtocol_Tests.swift
@@ -1,0 +1,10 @@
+
+import Foundation
+
+class Environment_Tests: XCTestCase {
+    
+    func dataForRequest_returnsAndResponseOnSuccess() async {
+        
+        
+    }
+}
